[test] WebAssembly itself in `workers` dir
